tension-app-server/requirements.txt
alembic==1.3.2
Bootstrap-Flask==1.2.0
Click==7.0
dominate==2.4.0
flasgger==0.9.3
Flask==1.1.1
Flask-JWT-Extended==3.24.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==2.6.0
Mako==1.1.0
MarkupSafe==1.1.1
mistune==0.8.4
PyJWT==1.7.1
PyMySQL==0.9.3
PyRSMQ==0.4.1
python-dateutil==2.8.1
python-editor==1.0.4
PyYAML==5.3
redis==3.4.1
six==1.14.0
SQLAlchemy==1.3.12
visitor==0.1.3
Werkzeug==0.16.0
